cleanup: simpfliy content system
